% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findErrorsInLine}
\alias{findErrorsInLine}
\title{Identifie les erreurs dans une ligne ou colonne}
\usage{
findErrorsInLine(line)
}
\arguments{
\item{line_data}{Vecteur représentant une ligne ou colonne}
}
\value{
Vecteur de booléens indiquant les cellules en erreur
}
\description{
Identifie les erreurs dans une ligne ou colonne
}
\keyword{internal}
