% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateTakuzu}
\alias{generateTakuzu}
\title{Génère une grille de Takuzu}
\usage{
generateTakuzu(size, fillRate)
}
\arguments{
\item{size}{La taille de la grille (nombre de lignes et de colonnes)}

\item{fillRate}{Le taux de remplissage de la grille (entre 0 et 1)}
}
\value{
Une matrice d'entiers représentant la grille de Takuzu générée
}
\description{
Crée une grille de Takuzu de taille spécifiée avec un taux de remplissage défini
}
\details{
Cette fonction génère une grille de Takuzu partiellement remplie en respectant
les règles du jeu : pas plus de deux 0 ou deux 1 consécutifs, et autant de 0 que de 1
dans chaque ligne et colonne. Les cellules non remplies sont marquées par -1.
}
