% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{solveTakuzu}
\alias{solveTakuzu}
\title{Résout une grille de Takuzu}
\usage{
solveTakuzu(grid)
}
\arguments{
\item{grid}{Grille initiale à résoudre}
}
\value{
Une matrice représentant la solution, ou NULL si pas de solution
}
\description{
Résout une grille de Takuzu
}
\keyword{internal}
