% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{takuzu_ui}
\alias{takuzu_ui}
\title{Interface utilisateur pour l'application Takuzu}
\usage{
takuzu_ui()
}
\value{
Un objet Shiny UI
}
\description{
Cette fonction génère l'interface utilisateur complète pour le jeu Takuzu, comprenant
une page d'accueil avec les règles et la sélection de difficulté, ainsi qu'une page de jeu
avec une grille interactive et des contrôles pour vérifier, obtenir des indices et afficher la solution.
}
\examples{
\dontrun{
# Utilisation avec shinyApp
shinyApp(ui = takuzu_ui(), server = takuzu_server)
}
}
